refactor: move curve transforms into new src/curves/ module
Extract cubic_to_quad, quad_to_cubic, merge_curves, and their shared
utilities (Cubic, TOLERANCE, split_cubic_at, cubic_farthest_fit_inside)
from src/transform/ and src/skia/ into a dedicated src/curves/ module.
src/transform/ now contains only render_bitmap and subpath operations;
src/skia/ is limited to skia-safe-dependent code.
