+++ src/transform/mod.rs
@@ -1,5 +1,2 @@
-pub(crate) mod cubic_to_quad;
-pub(crate) mod merge_curves;
-pub(crate) mod quad_to_cubic;
 pub(crate) mod render_bitmap;
 pub(crate) mod subpath;
